fix alcance de bombas

+++ model/bomb.go
@@ -6,7 +6,7 @@ const EXPLODE_TIME = 3
 
 type Bomb struct {
 	Position    Position
-	Alcance     int8
+	Alcance     int
 	PlantedTime time.Time
 	ExplodeTime time.Duration
 	Owner       *Player
@@ -16,7 +16,7 @@ func (b *Bomb) IsOwner(playerID string) bool {
 	return b.Owner.ID == playerID
 }
 
-func NewBomb(x, y float32, alcance int8, owner Player) *Bomb {
+func NewBomb(x, y float32, alcance int, owner Player) *Bomb {
 	return &Bomb{
 		Position:    Position{X: x, Y: y},
 		Alcance:     alcance,
